Show warning on delete sign in gloss edit view if references other glosses

+++ signbank/dictionary/update.py
@@ -34,6 +34,8 @@ from django.shortcuts import redirect
 from signbank.dictionary.update_senses_mapping import mapping_edit_keywords, mapping_group_keywords, mapping_add_keyword, \
     mapping_edit_senses_matrix, mapping_toggle_sense_tag
 from signbank.dictionary.consistency_senses import reorder_translations
+from signbank.dictionary.related_objects import gloss_related_objects
+
 
 def show_error(request, translated_message, form, dataset_languages):
     # this function is used by the add_gloss function below
@@ -740,9 +742,15 @@ def update_gloss(request, glossid):
         if value == 'confirmed':
             # delete the gloss and redirect back to gloss list
 
-            pk = gloss.pk
+            related_objects = gloss_related_objects(gloss)
+
+            if settings.GUARDED_GLOSS_DELETE and related_objects:
+                reverse_url = 'dictionary:admin_gloss_view'
+                messages.add_message(request, messages.INFO,
+                                     _("GUARDED_GLOSS_DELETE is set to True. The gloss has relations to other glosses and was not deleted."))
+                return HttpResponseRedirect(reverse(reverse_url, kwargs={'pk': gloss.id}))
+
             gloss.delete()
-            gloss.pk = pk
 
             return HttpResponseRedirect(reverse('dictionary:admin_gloss_list'))
